refactor: rename and reorganize routers, services, models, and storage for better clarity and separation of concerns
